Extract badge handling to a separate class
